# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/User/ChangePassword": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "ChangePassword a User",
                "operationId": "a09bed86d069b9d0cf0503e484022847",
                "parameters": [
                    {
                        "name": "now_password",
                        "in": "query",
                        "description": "User Now password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "new_password",
                        "in": "query",
                        "description": "User New password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "new_password_confirmation",
                        "in": "query",
                        "description": "User New password confirmation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User update successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/User/PersonalInformation": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "show a User",
                "operationId": "463ceb75866fd7a74a0ac12e5d0cac02",
                "parameters": [
                    {
                        "name": "firstName",
                        "in": "query",
                        "description": "firstName Passenger",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "description": "lastName Passenger",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nationalcode",
                        "in": "query",
                        "description": "nationalcode Passenger",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "gender Passenger (male,femail)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "birthday",
                        "in": "query",
                        "description": "birthday Passenger",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Passenger detail successfully"
                    },
                    "409": {
                        "description": "TrainTicket Validate Error"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/User/BankInformation": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Add bankInformation a User",
                "operationId": "9d34fb86956bd5a21db40d405db6e513",
                "parameters": [
                    {
                        "name": "shabanumber",
                        "in": "query",
                        "description": "User shabanumber",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cardnumber",
                        "in": "query",
                        "description": "User cardnumber",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add personalInformation successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/User/ChangeMobile": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Add mobile a User",
                "operationId": "3f57707fcc3aad620697ad1fc63435ee",
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "User mobile",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add personalInformation successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/User/ChangeEmail": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Add email a User",
                "operationId": "4f6cfd64d75c1623c8d4d18ea8cad9a6",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email mobile",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add personalInformation successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}